+++ app/core/Engine/controllers/transaction-controller/metrics_properties/hash.ts
