+++ app/core/Engine/controllers/transaction-controller/metrics_properties/metamask-pay.test.ts
@@ -62,11 +62,8 @@ describe('Metamask Pay Metrics', () => {
       allTransactions: [],
       getUIMetrics: getUIMetricsMock,
       getState: getStateMock,
-      transactionEventHandlerRequest: {
-        getState: getStateMock,
-        initMessenger: {} as never,
-        smartTransactionsController: {} as never,
-      },
+      initMessenger: {} as never,
+      smartTransactionsController: {} as never,
     };
   });
