+++ app/core/Engine/controllers/transaction-controller/metrics_properties/simulation-values.test.ts
@@ -29,11 +29,8 @@ describe('getSimulationValuesProperties', () => {
       allTransactions: [],
       getUIMetrics: getUIMetricsMock,
       getState: getStateMock,
-      transactionEventHandlerRequest: {
-        getState: getStateMock,
-        initMessenger: {} as never,
-        smartTransactionsController: {} as never,
-      },
+      initMessenger: {} as never,
+      smartTransactionsController: {} as never,
     };
   });
