+++ app/core/Engine/controllers/transaction-controller/metrics_properties/simulation-values.ts
