+++ app/core/Engine/controllers/transaction-controller/utils.ts
@@ -1,5 +1,3 @@
-import type { TransactionMeta } from '@metamask/transaction-controller';
-import { merge } from 'lodash';
 import { createProjectLogger } from '@metamask/utils';
 
 import { MetricsEventBuilder } from '../../../Analytics/MetricsEventBuilder';
@@ -8,12 +6,6 @@ import {
   IMetaMetricsEvent,
 } from '../../../Analytics/MetaMetrics.types';
 import { TRANSACTION_EVENTS } from '../../../Analytics/events/confirmations';
-import type {
-  TransactionEventHandlerRequest,
-  TransactionMetrics,
-  TransactionMetricsBuilder,
-} from './types';
-import { EMPTY_METRICS } from './constants';
 import Engine from '../../Engine';
 
 const log = createProjectLogger('transaction-metrics');
@@ -22,14 +14,6 @@ export function isFinalizedEvent(eventType: IMetaMetricsEvent): boolean {
   return (
     eventType.category === TRANSACTION_EVENTS.TRANSACTION_FINALIZED.category
   );
-}
-
-export function getConfirmationMetrics(
-  state: ReturnType<TransactionEventHandlerRequest['getState']>,
-  transactionId: string,
-): TransactionMetrics {
-  return (state?.confirmationMetrics?.metricsById?.[transactionId] ||
-    {}) as unknown as TransactionMetrics;
 }
 
 export function generateEvent({
@@ -61,53 +45,4 @@ export function retryIfEngineNotInitialized(fn: () => void): boolean {
 
     return true;
   }
-}
-
-export async function getBuilderMetrics({
-  builders,
-  eventType,
-  request,
-  transactionMeta,
-}: {
-  builders: TransactionMetricsBuilder[];
-  eventType: IMetaMetricsEvent;
-  request: TransactionEventHandlerRequest;
-  transactionMeta: TransactionMeta;
-}) {
-  const metrics = {
-    properties: {},
-    sensitiveProperties: {},
-  };
-
-  const allTransactions =
-    request.getState()?.engine?.backgroundState?.TransactionController
-      ?.transactions ?? [];
-
-  const getState = request.getState;
-
-  const getUIMetrics = (transactionId: string): TransactionMetrics =>
-    getConfirmationMetrics(getState(), transactionId);
-
-  const builderResults = await Promise.all(
-    builders.map(async (builder) => {
-      try {
-        return await builder({
-          eventType,
-          transactionMeta,
-          allTransactions,
-          getUIMetrics,
-          getState,
-          transactionEventHandlerRequest: request,
-        });
-      } catch (error) {
-        return EMPTY_METRICS;
-      }
-    }),
-  );
-
-  for (const currentMetrics of builderResults) {
-    merge(metrics, currentMetrics);
-  }
-
-  return metrics;
 }
